requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.9.0
transformers>=4.20.0

# Vietnamese NLP
underthesea>=1.3.0
pyvi>=0.1.1

# Visualization (MATPLOTLIB)
matplotlib>=3.5.0
seaborn>=0.11.0

# Dashboard
streamlit>=1.20.0

# API
flask>=2.0.0
flask-cors>=3.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0

# Huggingface Datasets
datasets>=2.0.0
